requirements.txt
flask
flask-cors
pymongo
ultralytics
pillow
